% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_utils.R
\name{compute_cumulative_elasticity}
\alias{compute_cumulative_elasticity}
\title{Compute cumulative elasticity from original and vaccinated NGMs}
\usage{
compute_cumulative_elasticity(ngm, ngm_vac, M)
}
\arguments{
\item{ngm}{matrix\link{2M,2M} - original NGM}

\item{ngm_vac}{matrix\link{2M,2M} - vaccinated NGM}

\item{M}{int - the number of age groups}
}
\value{
A numeric vector of cumulative elasticities by age group
}
\description{
Compute cumulative elasticity from original and vaccinated NGMs
}
