% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_utils.R
\name{compute_ngm_at_time}
\alias{compute_ngm_at_time}
\title{Compute the next-generation matrix (NGM) at a given time}
\usage{
compute_ngm_at_time(
  model_output,
  time_point,
  age_distribution_sa,
  n_active,
  D_max,
  ...
)
}
\arguments{
\item{model_output}{A data frame output from model_mpox_network}

\item{time_point}{Time at which to compute the NGM}

\item{age_distribution_sa}{Age distribution among sexually active groups}

\item{n_active}{Number of sexually active age groups}

\item{D_max}{Maximum degree}

\item{...}{Additional parameters passed to time_varying_ngm}
}
\value{
A list with \code{ngm} and the extracted \code{x_state}
}
\description{
Compute the next-generation matrix (NGM) at a given time
}
