% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_utils.R
\name{compute_S_in}
\alias{compute_S_in}
\title{Compute S_in matrix from xbar and age distribution}
\usage{
compute_S_in(xbar, age_distribution_sa)
}
\arguments{
\item{xbar}{A 3D array \link{i, j, n} representing the susceptible binding sites}

\item{age_distribution_sa}{A numeric vector for sexually active age distribution}
}
\value{
A matrix \link{i, n} where each element is the age-weighted sum over j
}
\description{
Compute S_in matrix from xbar and age distribution
}
