% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_utils.R
\name{compute_S_in_powered}
\alias{compute_S_in_powered}
\title{Raise S_in matrix elements to power n (column index)}
\usage{
compute_S_in_powered(S_in)
}
\arguments{
\item{S_in}{A matrix \link{i, n} representing weighted susceptible binding sites}
}
\value{
A matrix \link{i, n} with each element raised to the power of its column index
}
\description{
Raise S_in matrix elements to power n (column index)
}
