% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_utils.R
\name{compute_vaccinated_ngm}
\alias{compute_vaccinated_ngm}
\title{Compute vaccinated NGM given a S_in matrix and state}
\usage{
compute_vaccinated_ngm(
  S_in_powered,
  x_state,
  target = "random",
  deltaU = 1,
  gamma = gamma,
  beta = beta
)
}
\arguments{
\item{S_in_powered}{A matrix \link{i, n} of S_in raised to degree powers}

\item{x_state}{A list containing S, xS, xI, xbar}

\item{target}{default is "random"}

\item{deltaU}{default is 1}
}
\value{
A named list returned by time_varying_ngm
}
\description{
Compute vaccinated NGM given a S_in matrix and state
}
