% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_home}
\alias{contact_home}
\title{Synthetic contact matrices for the home location for 177 geographical locations}
\format{
A named list of 177 country codes, 16 by 16 numerical matrices.
}
\source{
Prem K, Zandvoort Kv, Klepac P, Eggo RM, Davies NG, et al. (2024)
Projecting contact matrices in 177 geographical regions: An update and comparison
with empirical data for the COVID-19 era. PLOS Computational Biology 20(9): e1012454.
\url{https://doi.org/10.1371/journal.pcbi.1012454}
}
\usage{
contact_home
}
\description{
This object stores the synthetic contact matrices for at-home contacts for 177
geographical locations
}
\keyword{datasets}
