% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_distribution_moments.R
\name{degree_distribution_moments}
\alias{degree_distribution_moments}
\title{First and second moment of a degree distribution}
\usage{
degree_distribution_moments(degree_distribution)
}
\arguments{
\item{degree_distribution}{vector of probabilities, where the index represents
the degree}
}
\value{
list of \code{first_moment} and \code{second_moment} of the degree distribution
}
\description{
First and second moment of a degree distribution
}
