% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_newly_infected_individual.R
\name{degree_newly_infected_individual}
\alias{degree_newly_infected_individual}
\title{Degree distribution of a newly infected individual as a function of time}
\usage{
degree_newly_infected_individual(
  n_active,
  D_max,
  degree_distribution,
  degree_sizebiased,
  age_distribution_sa,
  xI,
  xbar
)
}
\arguments{
\item{n_active}{int - number of sexually active age groups}

\item{D_max}{int - maximum degree in the population}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{degree_sizebiased}{vector - size biased degree distribution, calculated from \code{degree_distribution}}

\item{age_distribution_sa}{vector - age distribution of sexually active age groups, calculated from \code{age_distribution}}

\item{xI}{array of dimension \code{n_active}, \code{n_active}, \code{D_max} of susceptible binding sites with infectious partner}

\item{xbar}{array of dimension \code{n_active}, \code{n_active}, \code{D_max} of susceptible binding sites}
}
\value{
matrix
}
\description{
Degree distribution of a newly infected individual as a function of time
}
