% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_transmission_parameters.R
\name{determine_transmission_parameters}
\alias{determine_transmission_parameters}
\title{Determine transmission parameters from SAR, age distribution, contact matrix and recovery rate}
\usage{
determine_transmission_parameters(age_distribution, contact_matrix, gamma)
}
\arguments{
\item{age_distribution}{vector\link{M} - population age distribution}

\item{contact_matrix}{matrix\link{M, M} - global contact rates, contact_matrix\link{i, j} = contact_matrix\link{j, i}
per-pair global contact rate between age groups i and j}

\item{gamma}{double >= 0 - recovery rate per day}
}
\value{
list of transmission parameters
\itemize{
\item \code{p_cont} - transmission probability in global transmission
\item \code{c_ave} - average contact rate
\item \code{beta} - transmission rate sexual contacts
}
}
\description{
Determine transmission parameters from SAR, age distribution, contact matrix and recovery rate
}
