% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRC_pop2024}
\alias{DRC_pop2024}
\title{Population age distribution for the DRC}
\format{
tibble with 23 rows and three columns: age_group, male, female
}
\source{
United Nations, Department of Economic and Social Affairs, Population Division (2024).
World Population Prospects 2024, Online Edition. Copyright © 2024 by United Nations,
made available under a Creative Commons license CC BY 3.0 IGO
\url{https://population.un.org/wpp/}
}
\usage{
DRC_pop2024
}
\description{
Population age distriution for the DRC in age groups of five years with last
age group being 100+ years
}
\keyword{datasets}
