% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_vaccov_matrix.R
\name{eff_vaccov_matrix}
\alias{eff_vaccov_matrix}
\title{Effective vaccination coverage for age group i with degree n}
\usage{
eff_vaccov_matrix(
  deltaU = 1,
  VE = 0.8,
  S_in,
  age_distribution,
  M,
  N_1,
  N_2,
  population_size,
  degree_distribution,
  target_D = NULL,
  target = c("random", "age", "age-degree")
)
}
\arguments{
\item{deltaU}{double - unit change in the number of vaccines}

\item{VE}{double - vaccine efficacy}

\item{S_in}{matrix\link{n_active, D_max} - proportion susceptible in group (i,n) (the group in age i with degree n)}

\item{age_distribution}{vector\link{M} - population age distribution}

\item{M}{int - total number of age groups}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}

\item{population_size}{double - population size}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{target_D}{double - integer, minimum degree class to target with vaccination (used in "age-degree" targeting)}

\item{target}{character - targeting strategy; one of "random", "age", "age-degree"}
}
\value{
list with two elements
\itemize{
\item effV_mat = matrix\link{n_active, D_max} - effective vaccine coverage for group (i,n)
\item effV_age = vector\link{M} - effective vaccine coverage for age group i (all age groups 1,..,M)
}
}
\description{
Effective vaccination coverage for age group i with degree n
}
