% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_functions.R
\name{elas_func}
\alias{elas_func}
\title{Elasticity}
\usage{
elas_func(matA, delta_matA)
}
\arguments{
\item{matA}{original matrix A}

\item{delta_matA}{matrix with perturbed elements}
}
\value{
matrix
}
\description{
Elasticity
}
\examples{
matA <- matrix(1:16, nrow = 4, ncol = 4)
delta_matA <- matrix(1e-06, nrow = 4, ncol = 4)
elas_func(matA, delta_matA) # elasticity
}
