% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_utils.R
\name{extract_x_arrays_at_time}
\alias{extract_x_arrays_at_time}
\title{Extract reshaped x variables from model output at a specific time point}
\usage{
extract_x_arrays_at_time(model_output, time_point, n_active, D_max)
}
\arguments{
\item{model_output}{A data frame output from model_mpox_network}

\item{time_point}{A numeric time at which to extract the state}

\item{n_active}{Number of sexually active age groups}

\item{D_max}{Maximum degree}
}
\value{
A list containing reshaped x variables (x0, xbar, xE, xI, xS, S)
}
\description{
Extract reshaped x variables from model output at a specific time point
}
