% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_infection_agegroup.R
\name{initialize_infection_agegroup}
\alias{initialize_infection_agegroup}
\title{Initialize infection in specified age group}
\usage{
initialize_infection_agegroup(
  M,
  n_active,
  D_max,
  seed_infected,
  degree_zero,
  initial_age_group,
  degree_distribution,
  degree_sizebiased
)
}
\arguments{
\item{M}{int - total number of age groups}

\item{n_active}{int - number of sexually active age groups}

\item{D_max}{int - maximum degree in the population}

\item{seed_infected}{double between 0 and 1 - initial infected}

\item{degree_zero}{double - fraction of individuals with degree zero}

\item{initial_age_group}{int - index of age group where infection is seeded}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{degree_sizebiased}{vector - size biased degree distribution, calculated from \code{degree_distribution}}
}
\value{
named vector
}
\description{
Initialize infection in specified age group
}
