% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpoxnetwork-package.R
\docType{package}
\name{mpoxnetwork-package}
\alias{mpoxnetwork}
\alias{mpoxnetwork-package}
\title{mpoxnetwork: Code accompanying the study by Miura and Leung et al. (2025), *Targeted intervention strategies for clade Ib mpox with multiple transmission*, _medRxiv_ (https://doi.org/xxx).}
\description{
The R package `mpoxnetwork` implements a two-level epidemic model that explicitly incorporates both sexual and non-sexual (household) transmission routes. The model is flexible, age-structured, and built around a sexual contact network informed by power-law degree distributions. The package supports simulation of epidemics, projection of reproduction numbers by mode of transmission, final epidemic sizes, and the age- and mode-specific contributions to transmission over time. Detailed methodology is provided in the manuscript.
}
\author{
\strong{Maintainer}: Fuminari Miura \email{fuminari.miura@rivm.nl} (\href{https://orcid.org/0000-0001-8235-757X}{ORCID})

Authors:
\itemize{
  \item Ka Yin Leung \email{ka.yin.leung@rivm.nl} (\href{https://orcid.org/0000-0002-7596-7641}{ORCID})
}

}
\keyword{internal}
