% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGM_fn.R
\name{NGM_fn}
\alias{NGM_fn}
\title{Large age-structured next generation matrix - refer to Supplement 1.1.1}
\usage{
NGM_fn(
  beta,
  gamma,
  p_cont,
  degree_distribution,
  age_distribution,
  contact_matrix,
  M,
  N_1,
  N_2
)
}
\arguments{
\item{beta}{double >=0 - transmission rate sexual contact per day}

\item{gamma}{double >= 0 - recovery rate per day}

\item{p_cont}{double between 0 and 1 - transmission probability global contact}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{age_distribution}{vector\link{M} - population age distribution}

\item{contact_matrix}{matrix\link{M, M} - global contact rates, contact_matrix\link{i, j} = contact_matrix\link{j, i}
per-pair global contact rate between age groups i and j}

\item{M}{int - total number of age groups}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}
}
\value{
list
}
\description{
Large age-structured next generation matrix - refer to Supplement 1.1.1
}
