% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_data_Kami}
\alias{obs_data_Kami}
\title{Observed age distribution case data from Kamituga}
\format{
A data frame with age group and observed proportions.
}
\source{
Vakaniaki, E.H., Kacita, C., Kinganda-Lusamaki, E. et al.
Sustained human outbreak of a new MPXV clade I lineage in eastern
Democratic Republic of the Congo. Nat Med 30, 2791–2795 (2024).
\url{https://doi.org/10.1038/s41591-024-03130-3} licensed under a Creative Commons
Attribution 4.0 International License
}
\usage{
obs_data_Kami
}
\description{
This object stores the observed proportion of cases by age group
from the Kamituga dataset.
}
\keyword{datasets}
