% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mpox_network.R
\name{ode_model_mpox_network}
\alias{ode_model_mpox_network}
\title{ODE system for the mpox network model}
\usage{
ode_model_mpox_network(times, y, parms)
}
\arguments{
\item{times}{vector of times}

\item{y}{vector of variables}

\item{parms}{list of parameters}
}
\value{
list of solution to the ODE system over time
}
\description{
ODE system for the mpox network model
}
