% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_contact_matrix.R
\name{prepare_contact_matrix}
\alias{prepare_contact_matrix}
\title{Grouping of the oldest age groups into one age group in raw contact matrices for the DRC}
\usage{
prepare_contact_matrix(raw_contact_matrix, raw_age_dist, N_1, N_2, M)
}
\arguments{
\item{raw_contact_matrix}{matrix - raw contact matrix}

\item{raw_age_dist}{vector - age distribution corresponding to the age groups in the raw contact matrix}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}

\item{M}{int - total number of desired age groups}
}
\value{
matrix
}
\description{
Grouping of the oldest age groups into one age group in raw contact matrices for the DRC
}
