% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_incidence_data_for_plotting.R
\name{prepare_incidence_data_for_plotting}
\alias{prepare_incidence_data_for_plotting}
\title{Wrangles data from simulation into incidence data ready for plotting}
\usage{
prepare_incidence_data_for_plotting(
  data,
  population_size,
  age_distribution,
  N_1
)
}
\arguments{
\item{data}{tibble - simulation data of epidemic outbreak of the mpox network model}

\item{population_size}{int - total population size}

\item{age_distribution}{vector\link{dbl} - age distribution in the population, sums to one}

\item{N_1}{int - first sexually active age group}
}
\value{
tibble
}
\description{
Wrangles data from simulation into incidence data ready for plotting
}
