% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_susceptible_partner.R
\name{probability_susceptible_partner}
\alias{probability_susceptible_partner}
\title{The probability that a partner of a newly infected individual is susceptible}
\usage{
probability_susceptible_partner(
  n_active,
  D_max,
  age_distribution_sa,
  degree_sizebiased,
  xbar,
  xS
)
}
\arguments{
\item{n_active}{int - number of sexually active age groups}

\item{D_max}{int - maximum degree in the population}

\item{age_distribution_sa}{vector - age distribution of sexually active age groups, calculated from \code{age_distribution}}

\item{degree_sizebiased}{vector - size biased degree distribution, calculated from \code{degree_distribution}}

\item{xbar}{array of dimension \code{n_active}, \code{n_active}, \code{D_max} of susceptible binding sites}
}
\value{
matrix of dimension \code{n_active}, \code{D_max} with probabilities
}
\description{
refer to manuscript and supplement for explanation and derivation of this quantity
}
