% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process simulation data to population numbers in long format for further analyses}
\usage{
process_data(
  var_str,
  data,
  population_size,
  age_distribution,
  sexual_active_only = FALSE,
  N_1 = NULL
)
}
\arguments{
\item{var_str}{string - name of variable under consideration, e.g. S, E, I, R, xS, etc.}

\item{data}{tibble - epidemic outbreak data, output of \code{model_mpox_network}}

\item{population_size}{int - size of the total population}

\item{age_distribution}{vector\link{M} - population age distribution}

\item{sexual_active_only}{boolean - whether or variables are only applicable for sexually active age groups, default is FALSE}

\item{N_1}{int - first sexually active age group}
}
\value{
tibble - subset of data, modified to population numbers, in long format
}
\description{
Process simulation data to population numbers in long format for further analyses
}
