% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_incidence_data_for_plotting.R
\name{read_simulation_data}
\alias{read_simulation_data}
\title{Helper function to read RDS file containing simulation data}
\usage{
read_simulation_data(scenario_name, data_dir)
}
\arguments{
\item{scenario_name}{string - name of the scenario to read in}

\item{data_dir}{path - path to the directory containing the simulation data}
}
\value{
tibble
}
\description{
Helper function to read RDS file containing simulation data
}
