% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction_number.R
\name{reproduction_number}
\alias{reproduction_number}
\title{The basic reproduction number for the mpox network model}
\usage{
reproduction_number(
  beta,
  gamma,
  p_cont,
  degree_distribution,
  age_distribution,
  contact_matrix,
  M,
  N_1,
  N_2
)
}
\arguments{
\item{beta}{double >=0 - transmission rate sexual contact per day}

\item{gamma}{double >= 0 - recovery rate per day}

\item{p_cont}{double between 0 and 1 - transmission probability global contact}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{age_distribution}{vector\link{M} - population age distribution}

\item{contact_matrix}{matrix\link{M, M} - global contact rates, contact_matrix\link{i, j} = contact_matrix\link{j, i}
per-pair global contact rate between age groups i and j}

\item{M}{int - total number of age groups}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}
}
\value{
list with the following list elements
\itemize{
\item reproduction_number - the basic reproduction number
\item R_ss - the expected number of secondary cases from sexual to sexual transmission
\item R_gg - the expected number of secondary cases from global to global transmission
}
}
\description{
The basic reproduction number for the mpox network model
}
\examples{
reproduction_number(
  beta = 0.2, gamma = 1 / 5, p_cont = 0.05,
  degree_distribution = c(0.1, 0.3, 0.5, 0.05),
  age_distribution = c(0.1, 0.9),
  contact_matrix = matrix(1, nrow = 2, ncol = 2),
  M = 2, N_1 = 1, N_2 = 1
)
}
