% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lhs_ngm_simulation.R
\name{run_lhs_ngm_simulation}
\alias{run_lhs_ngm_simulation}
\title{Run LHS sampling and compute NGM outcomes}
\usage{
run_lhs_ngm_simulation(
  n_samples,
  scaled_samples,
  c_ave,
  age_distribution,
  contact_matrix,
  M,
  N_1,
  N_2
)
}
\arguments{
\item{n_samples}{Integer. Number of samples to simulate.}

\item{scaled_samples}{A numeric matrix of LHS samples (scaled to param ranges).}

\item{c_ave}{Double. Average household contact rate.}

\item{age_distribution}{Numeric vector. Population age distribution.}

\item{contact_matrix}{Numeric matrix. Age-specific contact matrix.}

\item{M}{Integer. Number of age groups.}

\item{N_1}{Integer. Index of first sexually active age group.}

\item{N_2}{Integer. Index of last sexually active age group.}
}
\value{
A data frame of simulation outcomes by sample and age group.
}
\description{
Run LHS sampling and compute NGM outcomes
}
