% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_functions.R
\name{sens_func}
\alias{sens_func}
\title{Sensitivity}
\usage{
sens_func(matA, delta_matA)
}
\arguments{
\item{matA}{original matrix A}

\item{delta_matA}{matrix with perturbed elements}
}
\value{
matrix
}
\description{
Sensitivity
}
\examples{
matA <- matrix(1:16, nrow = 4, ncol = 4)
delta_matA <- matrix(1e-06, nrow = 4, ncol = 4)
sens_func(matA, delta_matA)
}
