% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.R
\name{simulate_model}
\alias{simulate_model}
\title{Wrapper function to simulate epidemic outbreak for the mpox network model}
\usage{
simulate_model(
  M,
  N_1,
  N_2,
  contact_matrix,
  age_distribution,
  degree_zero,
  degree_distribution,
  beta,
  p_cont,
  sigma,
  gamma,
  seed_infected,
  time_end,
  time_increment,
  data_dir,
  data_name,
  initial,
  x_output = FALSE
)
}
\arguments{
\item{M}{int - total number of age groups}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}

\item{contact_matrix}{matrix\link{M, M} - global contact rates, contact_matrix\link{i, j} = contact_matrix\link{j, i}
per-pair global contact rate between age groups i and j}

\item{age_distribution}{vector\link{M} - population age distribution}

\item{degree_zero}{double >=0 - fraction of sexually active population with degree zero}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{beta}{double >=0 - transmission rate sexual contact per day}

\item{p_cont}{double between 0 and 1 - transmission probability global contact}

\item{sigma}{double >= 0 - 1/sigma mean latent period in days}

\item{gamma}{double >= 0 - recovery rate per day}

\item{seed_infected}{double between 0 and 1 - initial infected}

\item{time_end}{int - end time of simulation in days}

\item{time_increment}{double - increment time steps in simulation}

\item{data_dir}{path - user-provided data directory where simulation result is saved}

\item{data_name}{string - user-provided name for the simulation}

\item{initial}{vector - user provided initial value condition,
default is NULL corresponding to uniform distribution of infection over the age groups}

\item{x_output}{boolean - whether or not to save x-variables to the results}
}
\value{
tibble
}
\description{
Wrapper function to simulate epidemic outbreak for the mpox network model
}
