% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_varying_ngm.R
\name{time_varying_ngm}
\alias{time_varying_ngm}
\title{Function to compute time varying NGMs}
\usage{
time_varying_ngm(
  beta,
  p_cont,
  gamma,
  contact_matrix,
  age_distribution,
  age_distribution_sa,
  n_active,
  degree_distribution,
  degree_sizebiased,
  D_max,
  M,
  N_1,
  N_2,
  S,
  xS,
  xI,
  xbar
)
}
\arguments{
\item{beta}{double >=0 - transmission rate sexual contact per day}

\item{p_cont}{double between 0 and 1 - transmission probability global contact}

\item{gamma}{double >= 0 - recovery rate per day}

\item{contact_matrix}{matrix\link{M, M} - global contact rates, contact_matrix\link{i, j} = contact_matrix\link{j, i}
per-pair global contact rate between age groups i and j}

\item{age_distribution}{vector\link{M} - population age distribution}

\item{age_distribution_sa}{vector - age distribution of sexually active age groups, calculated from \code{age_distribution}}

\item{n_active}{int - number of sexually active age groups}

\item{degree_distribution}{vector - defective distribution of degrees 1,..., D_max, excluding degree_zero}

\item{degree_sizebiased}{vector - size biased degree distribution, calculated from \code{degree_distribution}}

\item{D_max}{int - maximum degree in the population}

\item{M}{int - total number of age groups}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}

\item{S}{vector of length \code{M} of susceptible fractions by age group}

\item{xI}{array of dimension \code{n_active}, \code{n_active}, \code{D_max} of susceptible binding sites with infectious partner}

\item{xbar}{array of dimension \code{n_active}, \code{n_active}, \code{D_max} of susceptible binding sites}

\item{xE}{array of dimension \code{n_active}, \code{n_active}, \code{D_max} of susceptible binding sites with exposed partner}
}
\value{
list with the following elements
\itemize{
\item \code{R_all_t} - time-varying reproduction number
\item \code{R_gg_t} - time-varying secondary cases from global to global transmission
\item \code{R_sg_t} - time-varying secondary cases from global to sexual transmission
\item \code{R_gs_t} - time-varying secondary cases from sexual to global transmission
\item \code{R_ss_t} - time-varying secondary cases from sexual to sexual transmission
\item \code{k_gg_large} - submatrix of the time-varying NGM consisting of secondary cases from global to global transmission
\item \code{k_sg_large} - submatrix of the time-varying NGM consisting of secondary cases from global to sexual transmission
\item \code{k_gs_large} - submatrix of the time-varying NGM consisting of secondary cases from sexual to global transmission
\item \code{k_ss_large} - submatrix of the time-varying NGM consisting of secondary cases from sexual to sexual transmission
\item \code{global_NGM_large} - time-varying NGM combining the four submatrices
}
}
\description{
refer to manuscript and supplement for details on derivation and equations
}
