% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccinated_sus_matrix.R
\name{vaccinated_sus_matrix}
\alias{vaccinated_sus_matrix}
\title{Function to compute effective vaccination coverage for age group i with degree n}
\usage{
vaccinated_sus_matrix(M, N_1, N_2, effV_mat, effV_age, S, xS, xI, xbar)
}
\arguments{
\item{M}{int - total number of age groups}

\item{N_1}{int - first sexually active age group}

\item{N_2}{int - last sexually active age group, 1 <= N_1 <= N_2 <= M}

\item{effV_mat}{matrix\link{n_active, D_max} - effective vaccine coverage for group (i,n)}

\item{effV_age}{vector\link{M} - effective vaccine coverage for all age groups}

\item{S}{vector\link{M} - proportion susceptible for all age groups}

\item{xS}{array\link{n_active, n_active, D_max, D_max} - probability that a binding site is with group (i,n) and having a susceptible partner of group (j,k). Array index = \link{i, j, n, k}}

\item{xI}{array\link{n_active, n_active, D_max} - probability that a binding site is with group (i,n) and having an infectious partner of group j. Array index = \link{i, j, n}}

\item{xbar}{array\link{n_active, n_active, D_max} - probability that a binding site is with group (i,n) and having a partner of group j. Array index = \link{i, j, n}}
}
\value{
list with four elements
\itemize{
\item xS_vac array\link{n_active, n_active, D_max, D_max} - xS after vaccination
\item xI_vac array\link{n_active, n_active, D_max} - xI after vaccination
\item xbar_vac array\link{n_active, n_active, D_max} - xbar after vaccination
\item S_vac vector\link{M} - susceptible proportion after vaccination for all age groups
}
}
\description{
Function to compute effective vaccination coverage for age group i with degree n
}
